services/kafka_to_feature_store/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["gerard-martinez <gerard.martinez@king.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
quixstreams = "^2.5.1"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.3.1"
loguru = "^0.7.2"
hopsworks = "^3.7.0"


[tool.ruff]
line-length = 88

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true

[tool.ruff.lint]
extend-select = ["I"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
